requirements.txt
fastapi~=0.82.0
uvicorn[standard]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
torch~=1.12.0
transformers~=4.22.1
numpy~=1.21.4
Pillow~=9.2.0
matplotlib~=3.5.2
torchvision~=0.13.1
ray~=2.0.0